@@ -4,80 +4,88 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start-uhero-staging": "ng serve --app=uhero --env=staging --port=4200",
-    "start-coh-staging": "ng serve --app=coh --env=staging --port=4200",
-    "start-nta-staging": "ng serve --app=nta --env=staging --port=4200",
-    "start-uhero-dev": "ng serve --app=uhero --env=dev",
-    "start-coh-dev": "ng serve --app=coh --env=dev",
-    "start-nta-dev": "ng serve --app=nta --env=dev",
-    "start-uhero-api-dev": "ng serve --app=uhero --env=apiDev",
-    "start-coh-api-dev": "ng serve --app=coh --env=apiDev",
-    "start-nta-api-dev": "ng serve --app=nta --env=apiDev",
-    "build-uhero-staging": "ng build --app=uhero --env=staging --output-path=dist_staging",
-    "build-coh-staging": "ng build --app=coh --env=staging --base-href=/coh/ --output-path=cohDist_staging",
-    "build-nta-staging": "ng build --app=nta --env=staging --base-href=/nta/ --output-path=ntaDist_staging",
-    "build-uhero": "ng build --app=uhero --prod",
-    "build-coh": "ng build --app=coh --prod --base-href=/coh/",
-    "build-nta": "ng build --app=nta --prod --base-href=/nta/",
+    "start-uhero-staging": "ng serve --project=uhero -c=staging --port=4200",
+    "start-coh-staging": "ng serve --project=coh -c=staging --port=4200",
+    "start-nta-staging": "ng serve --project=nta -c=staging --port=4200",
+    "start-uhero-dev": "ng serve --project=uhero",
+    "start-coh-dev": "ng serve --project=coh",
+    "start-nta-dev": "ng serve --project=nta",
+    "start-uhero-api-dev": "ng serve --project=uhero -c=apiDev",
+    "start-coh-api-dev": "ng serve --project=coh -c=apiDev",
+    "start-nta-api-dev": "ng serve --project=nta -c=apiDev",
+    "build-uhero-staging": "ng build --project=uhero -c=staging --output-path=dist_staging",
+    "build-coh-staging": "ng build --project=coh -c=staging --base-href=/coh/ --output-path=cohDist_staging",
+    "build-nta-staging": "ng build --project=nta -c=staging --base-href=/nta/ --output-path=ntaDist_staging",
+    "build-uhero": "ng build --project=uhero --prod",
+    "build-coh": "ng build --project=coh --prod --base-href=/coh/",
+    "build-nta": "ng build --project=nta --prod --base-href=/nta/",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "engines": {
     "node": "8.9.4",
-    "npm": "5.8.0"
+    "npm": "6.2.0"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/router": "^5.2.9",
-    "angular2-highcharts": "https://github.com/gevgeny/angular2-highcharts",
+    "@angular/animations": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/platform-server": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "bootstrap": "^4.0.0-beta.2",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.4",
     "datatables.net": "^1.10.13",
     "datatables.net-buttons": "^1.2.4",
     "datatables.net-buttons-dt": "^1.2.4",
     "datatables.net-dt": "^1.10.13",
     "datatables.net-fixedcolumns": "^3.2.2",
     "datatables.net-fixedcolumns-dt": "^3.2.2",
+    "highcharts": "^6.1.1",
+    "highcharts-angular": "^2.0.1",
     "highcharts-export-csv": "^1.4.3",
     "ion-rangeslider": "^2.1.7",
-    "ng-recaptcha": "^3.0.3",
+    "ng-recaptcha": "4.0.0-beta.1",
     "popper.js": "^1.12.9",
-    "primeng": "4.2.2",
-    "rxjs": "^5.5.6",
+    "primeng": "^6.0.0-beta.1",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
     "tether": "^1.4.0",
     "tlite": "^0.1.6",
-    "zone.js": "^0.8.19",
-    "@angular/platform-server": "^5.2.9"
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.9",
+    "@angular-devkit/build-angular": "0.7.0-rc.3",
+    "@angular-devkit/architect": "0.7.0-rc.2",
+    "@angular-devkit/core": "0.7.0-rc.2",
+    "@angular-devkit/schematics": "0.7.0-rc.2",
+    "hoek": "5.0.3",
+    "@angular/cli": "6.2.0-beta.0",
+    "@angular/compiler-cli": "^6.1.0",
     "@types/google.analytics": "0.0.33",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
+    "@types/highcharts": "^5.0.25",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^2.0.41",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@types/node": "~10.5.2",
+    "codelyzer": "~4.4.2",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~2.0.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.6.2"
+    "protractor": "~5.3.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.10.0",
+    "typescript": "2.7.2"
   }
 }
