@@ -2,22 +2,7 @@ import { BrowserModule, Title } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { ChartModule } from 'angular2-highcharts';
-// Temp workaround for build errors
-// See: https://github.com/gevgeny/angular2-highcharts/issues/160
-import { HighchartsStatic } from 'angular2-highcharts/dist/HighchartsService';
-declare var require: any;
-export function highchartsFactory() {
-  const highcharts = require('highcharts/js/highstock');
-  const exp = require('highcharts/js/modules/exporting');
-  const offlineExport = require('highcharts/js/modules/offline-exporting');
-  const csv = require('./csv-export');
-
-  exp(highcharts);
-  offlineExport(highcharts);
-  csv(highcharts);
-  return (highcharts);
-}
+import { HighchartsChartModule } from 'highcharts-angular';
 import { Shared } from './shared/shared.module';
 import { routing } from './app.routes';
 import { AppComponent } from './app.component';
@@ -42,7 +27,7 @@ import { RouterModule } from '@angular/router';
     Shared,
     routing,
     BrowserModule.withServerTransition({ appId: 'serverApp' }),
-    ChartModule,
+    HighchartsChartModule,
     FormsModule,
     ReactiveFormsModule,
     HttpModule,
@@ -63,10 +48,6 @@ import { RouterModule } from '@angular/router';
     AnalyzerService,
     Title,
     ClipboardService,
-    {
-      provide: HighchartsStatic,
-      useFactory: highchartsFactory
-    },
     {
       provide: 'rootCategory',
       useValue: 59
