@@ -3,27 +3,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { ChartModule } from 'angular2-highcharts';
-// Temp workaround for build errors
-// See: https://github.com/gevgeny/angular2-highcharts/issues/160
-import { HighchartsStatic } from 'angular2-highcharts/dist/HighchartsService';
-declare var require: any;
-export function highchartsFactory() {
-  const highcharts = require('highcharts/js/highstock');
-  const exp = require('highcharts/js/modules/exporting');
-  const offlineExport = require('highcharts/js/modules/offline-exporting');
-  const csv = require('../csv-export');
-
-  exp(highcharts);
-  offlineExport(highcharts);
-  csv(highcharts);
-  highcharts.setOptions({
-    lang: {
-      thousandsSep: ','
-    }
-  });
-  return (highcharts);
-}
+import { HighchartsChartModule } from 'highcharts-angular';
 import { Shared } from '../shared/shared.module';
 import { routing } from '../nta.routes';
 import { DataTableModule, SharedModule } from 'primeng/primeng';
@@ -54,11 +34,12 @@ import { ClipboardService } from '../clipboard.service';
     Shared,
     routing,
     BrowserModule,
-    ChartModule,
     FormsModule,
     ReactiveFormsModule,
     HttpModule,
-    DataTableModule, SharedModule,
+    HighchartsChartModule,
+    DataTableModule,
+    SharedModule,
   ],
   entryComponents: [
     NtaHelpComponent
@@ -75,10 +56,6 @@ import { ClipboardService } from '../clipboard.service';
     HelpService,
     ClipboardService,
     Title,
-    {
-      provide: HighchartsStatic,
-      useFactory: highchartsFactory
-    },
     {
       provide: 'rootCategory',
       useValue: 2487
