@@ -1,6 +1,7 @@
 import { Component, Inject, OnInit, OnChanges, Input, ViewEncapsulation } from '@angular/core';
 import { HelperService } from '../helper.service';
-import * as Highcharts from 'highcharts';
+import * as Highcharts from 'highcharts/js/highcharts';
+import { HighchartsObject } from '../HighchartsObject';
 
 @Component({
   selector: 'app-highchart',
@@ -16,8 +17,9 @@ export class HighchartComponent implements OnInit, OnChanges {
   @Input() chartEnd;
   @Input() minValue;
   @Input() maxValue;
-  public options: Object;
-  private chart;
+  Highcharts = Highcharts;
+  chartOptions = <HighchartsObject>{};
+  updateChart = false;
 
   static findLastValue(valueArray) {
     let counter = valueArray.length - 1;
@@ -34,19 +36,22 @@ export class HighchartComponent implements OnInit, OnChanges {
 
   ngOnInit() {
     if (this.seriesData.seriesInfo === 'No data available' || this.seriesData.categoryDisplay.chartData.level.length === 0) {
+      this.updateChart = true;
       this.noDataChart(this.seriesData);
     } else {
+      this.updateChart = true;
       this.drawChart(this.seriesData, this.currentFreq, this.portalSettings, this.minValue, this.maxValue, this.chartStart, this.chartEnd);
     }
   }
 
   ngOnChanges() {
+    this.updateChart = true;
     this.drawChart(this.seriesData, this.currentFreq, this.portalSettings, this.minValue, this.maxValue, this.chartStart, this.chartEnd);
   }
 
-  getSeriesStartAndEnd(dates, start, end) {
+  getSeriesStartAndEnd = (dates, start, end) => {
     const defaultRanges = this._helper.setDefaultCategoryRange(this.currentFreq, dates, this.defaultRange);
-    let startIndex = defaultRanges.start, endIndex = defaultRanges.end;
+    let { startIndex, endIndex } = defaultRanges;
     dates.forEach((item, index) => {
       if (item.date === start) {
         startIndex = index;
@@ -55,33 +60,59 @@ export class HighchartComponent implements OnInit, OnChanges {
         endIndex = index;
       }
     });
-    return { start: startIndex, end: endIndex };
-  }
+    return { seriesStart: startIndex, seriesEnd: endIndex };
+  };
 
-  drawChart(seriesData, currentFreq, portalSettings, min, max, start?, end?) {
-    const dates = seriesData.categoryDisplay.chartData.dates;
-    const seriesRange = this.getSeriesStartAndEnd(dates, start, end);
-    const seriesStart = seriesRange.start;
-    const seriesEnd = seriesRange.end;
-    let series0 = seriesData.categoryDisplay.chartData[portalSettings.highcharts.series0Name];
-    let series1 = seriesData.categoryDisplay.chartData[portalSettings.highcharts.series1Name];
-    series0 = series0 ? series0.slice(seriesStart, seriesEnd + 1) : null;
-    series1 = series1 ? series1.slice(seriesStart, seriesEnd + 1) : null;
-    let startDate;
-    if (!start) {
-      startDate = dates[seriesStart].date;
-    }
-    // Check how many non-null points exist in level series
-    const pointCount = series0.filter(value => Number.isFinite(value));
+  setChartTitle = (title: string) => {
+    return {
+      text: title,
+      useHTML: true,
+      align: 'left',
+      widthAdjust: 0,
+      x: 0,
+      y: -5,
+      style: {
+        margin: 75
+      }
+    };
+  };
+
+  setXAxis = () => {
+    return {
+      type: 'datetime',
+      labels: {
+        enabled: false
+      },
+      lineWidth: 0,
+      tickLength: 0
+    };
+  };
+
+  setYAxis = (min?: number, max?: number) => {
+    return [{
+      labels: {
+        enabled: false
+      },
+      title: {
+        text: ''
+      },
+      minTickInterval: 0.01
+    }, {
+      labels: {
+        enabled: false
+      },
+      title: {
+        text: ''
+      },
+      min: min,
+      max: max,
+      minTickInterval: 0.01,
+      opposite: true
+    }];
+  };
+
+  setChartSeries = (portalSettings, series0, currentFreq, startDate, pseudoZones, series1) => {
     const chartSeries = [];
-    const minValue = min;
-    const maxValue = max;
-    const pseudoZones = seriesData.categoryDisplay.chartData.pseudoZones;
-    const decimals = seriesData.seriesInfo.decimals ? seriesData.seriesInfo.decimals : 1;
-    const percent = seriesData.seriesInfo.percent;
-    const title = seriesData.seriesInfo.displayName;
-    const dataFreq = currentFreq;
-    const unitsShort = seriesData.seriesInfo.unitsLabelShort;
     chartSeries.push({
       name: portalSettings.highcharts.series0Name,
       type: portalSettings.highcharts.series0Type,
@@ -89,7 +120,7 @@ export class HighchartComponent implements OnInit, OnChanges {
       data: series0,
       pointInterval: currentFreq === 'Q' ? 3 : currentFreq === 'S' ? 6 : 1,
       pointIntervalUnit: currentFreq === 'A' ? 'year' : 'month',
-      pointStart: start ? Date.parse(start) : Date.parse(startDate),
+      pointStart: startDate,
       states: {
         hover: {
           lineWidth: 2
@@ -109,265 +140,213 @@ export class HighchartComponent implements OnInit, OnChanges {
         data: series1,
         pointInterval: currentFreq === 'Q' ? 3 : currentFreq === 'S' ? 6 : 1,
         pointIntervalUnit: currentFreq === 'A' ? 'year' : 'month',
-        pointStart: start ? Date.parse(start) : Date.parse(startDate),
+        pointStart: startDate,
         dataGrouping: {
           enabled: false
         },
       });
     }
+    return chartSeries;
+  }
 
-    this.options = {
-      chart: {
-        spacingTop: 20, /* Add spacing to draw plot below fixed tooltip */
-        className: pointCount.length === 1 ? 'single-point' : undefined
-      },
-      exporting: {
-        enabled: false
-      },
-      title: {
-        text: '<br>',
-        useHTML: true,
-        align: 'left',
-        widthAdjust: 0,
-        x: 0,
-        y: -5,
-        style: {
-          margin: 75
-        }
-      },
-      tooltip: {
-        positioner: function () {
-          return { x: 0, y: -5 };
-        },
-        shadow: false,
-        borderWidth: 0,
-        shared: true,
-        formatter: function () {
-          const getLabelName = function (seriesName, freq, precent) {
-            if (seriesName === 'level') {
-              return ': ';
+  drawChart = (seriesData, currentFreq: string, portalSettings, min: number, max: number, chartStart?, chartEnd?) => {
+    const { dates, pseudoZones } = seriesData.categoryDisplay.chartData;
+    const { series0Name, series1Name } = portalSettings.highcharts;
+    const { start, end } = seriesData.categoryDisplay;
+    const { percent, title, unitsLabelShort, displayName } = seriesData.seriesInfo;
+    const { seriesStart, seriesEnd } = this.getSeriesStartAndEnd(dates, chartStart, chartEnd);
+    const decimals = seriesData.seriesInfo.decimals ? seriesData.seriesInfo.decimals : 1;
+    let series0 = seriesData.categoryDisplay.chartData[series0Name];
+    let series1 = seriesData.categoryDisplay.chartData[series1Name];
+    series0 = series0 ? series0.slice(seriesStart, seriesEnd + 1) : null;
+    series1 = series1 ? series1.slice(seriesStart, seriesEnd + 1) : null;
+    const startDate = chartStart ? Date.parse(chartStart) : Date.parse(dates[seriesStart].date);
+    // Check how many non-null points exist in level series
+    const levelLength = series0.filter(value => Number.isFinite(value));
+    const chartSeries = this.setChartSeries(portalSettings, series0, currentFreq, startDate, pseudoZones, series1);
+    const addSubtitle = (point0, freq, chart, point1?, series1?) => {
+      const dateLabel = this.formatDateLabel(point0.x, freq);
+      let subtitleText = '';
+      subtitleText += Highcharts.numberFormat(point0.y, decimals, '.', ',') + '<br> (' + unitsLabelShort + ') <br>';
+      subtitleText += series1 ?
+        this.formatTransformLabel(series1.name, percent) + '<br>' + Highcharts.numberFormat(point1.y, decimals, '.', ',') + '<br>' + dateLabel :
+        dateLabel;
+      chart.setSubtitle({
+        text: subtitleText,
+        verticalAlign: 'middle',
+        y: -20
+      });
+    }
+    const refreshTooltip = (chart, tooltipData, latest0, series0) => {
+      chart.tooltip.refresh(tooltipData);
+      series0.points[latest0].setState('');
+      const pointCount = series0.points.filter(point => Number.isFinite(point.y));
+      if (pointCount.length > 1) {
+        chart.setSubtitle({ text: '' });
+      }
+    };
+    const formatDate = (date, freq) => this._helper.formatDate(date, freq);
+    const checkPointCount = (freq, series0, point0, chart, point1?, series1?) => {
+      // Fiilter out null values
+      const pointCount = series0.points.filter(point => Number.isFinite(point.y));
+      // If only 1 non-null value exists, display data value as text
+      if (pointCount.length === 1) {
+        addSubtitle(point0, freq, chart, point1, series1);
+        series0.userOptions.states.hover.enabled = false;
+        series0.options.marker.states.hover.enabled = false;
+        point0.setState('');
+      }
+      if (pointCount.length > 1) {
+        chart.setSubtitle({ text: '' });
+      }
+    }
+    this.chartOptions.chart = {
+      spacingTop: 20,
+      className: levelLength.length === 1 ? 'single-point' : undefined,
+      events: {
+        render: function () {
+          const series0 = this.series[0];
+          const series1 = this.series[1];
+          // Get position of last non-null value
+          const latestSeries0 = (series0 !== undefined && series0.points && series0.points.length) ? HighchartComponent.findLastValue(series0.points) : -1;
+          const latestSeries1 = (series1 !== undefined && series1.points && series1.points.length) ? HighchartComponent.findLastValue(series1.points) : -1;
+
+          // Prevent tooltip from being hidden on mouseleave
+          // Reset toolip value and marker to most recent observation
+          this.tooltip.hide = () => {
+            if (latestSeries0 > -1 && latestSeries1 > -1) {
+              const tooltipData = [series0.points[latestSeries0], series1.points[latestSeries1]]
+              refreshTooltip(this, tooltipData, latestSeries0, series0);
             }
-            if (seriesName === 'c5ma') {
-              return percent ? 'Annual Chg: ' : 'Annual % Chg: ';
-            }
-            if (seriesName === 'ytd' && freq === 'A') {
-              return percent ? 'Year/Year Chg: ' : 'Year/Year % Chg: ';
-            }
-            if (seriesName === 'ytd' && freq !== 'A') {
-              return percent ? 'Year-to-Date Chg: ' : 'Year-to-Date % Chg: ';
+            // If there are no YTD values
+            if (latestSeries0 > -1 && latestSeries1 === -1) {
+              const tooltipData = [series0.points[latestSeries0]];
+              refreshTooltip(this, tooltipData, latestSeries0, series0);
             }
           };
-          const getFreqLabel = function (freq, date) {
-            if (freq === 'A') {
-              return '';
-            }
-            if (freq === 'Q') {
-              if (Highcharts.dateFormat('%b', date) === 'Jan') {
-                return 'Q1 ';
-              }
-              if (Highcharts.dateFormat('%b', date) === 'Apr') {
-                return 'Q2 ';
-              }
-              if (Highcharts.dateFormat('%b', date) === 'Jul') {
-                return 'Q3 ';
-              }
-              if (Highcharts.dateFormat('%b', date) === 'Oct') {
-                return 'Q4 ';
-              }
-            }
-            if (freq === 'M' || freq === 'S') {
-              return Highcharts.dateFormat('%b', date) + ' ';
-            }
-          };
-          const pseudo = 'Pseudo History ';
-          let s = '<b>' + title + '</b><br>';
-          if (pointCount.length > 1) {
-            // Get Quarter or Month for Q/M frequencies
-            s = s + getFreqLabel(dataFreq, this.x);
-            // Add year
-            s = s + Highcharts.dateFormat('%Y', this.x) + '';
-            this.points.forEach((point) => {
-              const displayValue = Highcharts.numberFormat(point.y, decimals, '.', ',');
-              const formattedValue = displayValue === '-0.00' ? '0.00' : displayValue;
-              const name = getLabelName(point.series.name, dataFreq, percent);
-              let label = name + formattedValue;
-              if (point.series.name === 'level') {
-                label += ' (' + unitsShort + ') <br>';
-              }
-              if (pseudoZones.length > 0) {
-                pseudoZones.forEach((zone) => {
-                  if (point.x < zone.value) {
-                    const otherSeriesLabel = pseudo + name + formattedValue;
-                    const levelLabel = otherSeriesLabel + ' (' + unitsShort + ') <br>';
-                    s += point.series.name === 'level' ? levelLabel : otherSeriesLabel;
-                    s += pseudo + name + formattedValue;
-                  }
-                  if (point.x >= zone.value) {
-                    s += label;
-                  }
-                });
-              }
-              if (pseudoZones.length === 0) {
-                s += label;
-              }
-            });
+          // Display tooltip when chart loads
+          if (latestSeries0 > -1 && latestSeries1 > -1) {
+            this.tooltip.refresh([series0.points[latestSeries0], series1.points[latestSeries1]]);
+            checkPointCount(currentFreq, series0, series0.points[latestSeries0], this, series1.points[latestSeries1], series1);
           }
+          if (latestSeries0 > -1 && latestSeries1 === -1) {
+            this.tooltip.refresh([series0.points[latestSeries0]]);
+            checkPointCount(currentFreq, series0, series0.points[latestSeries0], this);
+          }
+          // If no data available for a given date range, display series title and display dates where data is available for a series
+          if (latestSeries0 === -1 && latestSeries1 === -1) {
+            this.setClassName(undefined);
+            const categoryDisplayStart = formatDate(start, currentFreq);
+            const categoryDisplayEnd = formatDate(end, currentFreq);
+            this.setTitle({ text: '<b>' + title + '</b>' });
+            this.setSubtitle({ text: 'Data Available From: ' + categoryDisplayStart + ' - ' + categoryDisplayEnd, verticalAlign: 'middle', y: -20 });
+          }
+        }
+      }
+    };
+    this.chartOptions.exporting = { enabled: false };
+    this.chartOptions.title = this.setChartTitle('<br>');
+    this.chartOptions.tooltip = {
+      positioner: function () {
+        return { x: 0, y: -5 };
+      },
+      shadow: false,
+      borderWidth: 0,
+      shared: true,
+      formatter: function () {
+        const getLabelName = (seriesName, freq, percent) => {
+          if (seriesName === 'level') {
+            return ': ';
+          }
+          if (seriesName === 'c5ma') {
+            return percent ? 'Annual Chg: ' : 'Annual % Chg: ';
+          }
+          if (seriesName === 'ytd' && freq === 'A') {
+            return percent ? 'Year/Year Chg: ' : 'Year/Year % Chg: ';
+          }
+          if (seriesName === 'ytd' && freq !== 'A') {
+            return percent ? 'Year-to-Date Chg: ' : 'Year-to-Date % Chg: ';
+          }
+        };
+        const getFreqLabel = (freq, date) => {
+          if (freq === 'A') {
+            return '';
+          }
+          if (freq === 'Q') {
+            if (Highcharts.dateFormat('%b', date) === 'Jan') {
+              return 'Q1 ';
+            }
+            if (Highcharts.dateFormat('%b', date) === 'Apr') {
+              return 'Q2 ';
+            }
+            if (Highcharts.dateFormat('%b', date) === 'Jul') {
+              return 'Q3 ';
+            }
+            if (Highcharts.dateFormat('%b', date) === 'Oct') {
+              return 'Q4 ';
+            }
+          }
+          if (freq === 'M' || freq === 'S') {
+            return Highcharts.dateFormat('%b', date) + ' ';
+          }
+        };
+        const getSeriesLabel = (points, s) => {
+          points.forEach((point) => {
+            const displayValue = Highcharts.numberFormat(point.y, decimals, '.', ',');
+            const formattedValue = displayValue === '-0.00' ? '0.00' : displayValue;
+            const name = getLabelName(point.series.name, currentFreq, percent);
+            let label = name + formattedValue;
+            if (point.series.name === 'level') {
+              label += ' (' + unitsLabelShort + ') <br>';
+            }
+            if (pseudoZones.length) {
+              pseudoZones.forEach((zone) => {
+                if (point.x < zone.value) {
+                  const otherSeriesLabel = pseudo + name + formattedValue;
+                  const levelLabel = otherSeriesLabel + ' (' + unitsLabelShort + ') <br>';
+                  s += point.series.name === 'level' ? levelLabel : otherSeriesLabel;
+                  s += pseudo + name + formattedValue;
+                }
+                if (point.x >= zone.value) {
+                  s += label;
+                }
+              });
+            }
+            if (pseudoZones.length === 0){
+              s += label;
+            }
+          });
           return s;
-        },
-        useHTML: true
-      },
-      legend: {
-        enabled: false
-      },
-      credits: {
-        enabled: false
-      },
-      xAxis: {
-        type: 'datetime',
-        labels: {
-          enabled: false
-        },
-        lineWidth: 0,
-        tickLength: 0
-      },
-      yAxis: [{
-        labels: {
-          enabled: false
-        },
-        title: {
-          text: ''
-        },
-        minTickInterval: 0.01,
-      }, {
-        title: {
-          text: ''
-        },
-        labels: {
-          enabled: false
-        },
-        min: minValue,
-        max: maxValue,
-        minTickInterval: 0.01,
-        opposite: true
-      }],
-      plotOptions: {
-        line: {
-          marker: {
-            enabled: pointCount.length === 1 ? false : true,
-            radius: 1.5
-          }
         }
-      },
-      series: chartSeries,
-    };
-  }
-
-  noDataChart(seriesData) {
-    const title = seriesData.seriesInfo.title === undefined ? seriesData.seriesInfo.name : seriesData.seriesInfo.title;
-
-    this.options = {
-      title: {
-        text: '<b>' + title + '</b><br>' + 'No Data Available',
-        align: 'left',
-        widthAdjust: 0,
-      },
-      exporting: {
-        enabled: false
-      },
-      legend: {
-        enabled: false
-      },
-      credits: {
-        enabled: false
-      },
-      yAxis: [{
-        title: {
-          text: ''
+        const pseudo = 'Pseudo History ';
+        let s = '<b>' + displayName + '</b><br>';
+        if (levelLength.length > 1) {
+          // Get Quarter or Month for Q/M frequencies
+          s = s + getFreqLabel(currentFreq, this.x);
+          // Add year
+          s = s + Highcharts.dateFormat('%Y', this.x) + '';
+          s = getSeriesLabel(this.points, s);
         }
-      }],
-      xAxis: {
-        lineWidth: 0
+        return s;
       },
-      series: [{
-        data: []
-      }],
-      lang: {
-        noData: 'No Data Available'
+      useHTML: true
+    };
+    this.chartOptions.legend = { enabled: false };
+    this.chartOptions.credits = { enabled: false };
+    this.chartOptions.xAxis = this.setXAxis();
+    this.chartOptions.yAxis = this.setYAxis(min, max);
+    this.chartOptions.plotOptions = {
+      line: {
+        marker: {
+          enabled: levelLength.length === 1 ? false : true,
+          radius: 1.5
+        }
       }
     };
+    this.chartOptions.series = chartSeries;
   }
 
-  render(event) {
-    this.chart = event;
-    const series0 = this.chart.series[0];
-    const series1 = this.chart.series[1];
-    // Get position of last non-null value
-    const latestSeries0 = (series0 !== undefined && series0.points && series0.points.length) ? HighchartComponent.findLastValue(series0.points) : -1;
-    const latestSeries1 = (series1 !== undefined && series1.points && series1.points.length) ? HighchartComponent.findLastValue(series1.points) : -1;
-
-    // Prevent tooltip from being hidden on mouseleave
-    // Reset toolip value and marker to most recent observation
-    this.chart.tooltip.hide = function () {
-      if (latestSeries0 > -1 && latestSeries1 > -1) {
-        this.chart.tooltip.refresh([series0.points[latestSeries0], series1.points[latestSeries1]]);
-        series0.points[latestSeries0].setState('');
-      }
-      // If there are no YTD values
-      if (latestSeries0 > -1 && latestSeries1 === -1) {
-        this.chart.tooltip.refresh([series0.points[latestSeries0]]);
-        series0.points[latestSeries0].setState('');
-      }
-    };
-    // Display tooltip when chart loads
-    if (latestSeries0 > -1 && latestSeries1 > -1) {
-      this.chart.tooltip.refresh([series0.points[latestSeries0], series1.points[latestSeries1]]);
-      this.checkPointCount(series0, series0.points[latestSeries0], series1.points[latestSeries1], series1);
-    }
-    // If there are no YTD values
-    if (latestSeries0 > -1 && latestSeries1 === -1) {
-      this.chart.tooltip.refresh([series0.points[latestSeries0]]);
-      this.checkPointCount(series0, series0.points[latestSeries0]);
-    }
-    // If no data available for a given date range, display series title and display dates where data is available for a series
-    if (latestSeries0 === -1 && latestSeries1 === -1) {
-      this.chart.setClassName(undefined);
-      const start = this._helper.formatDate(this.seriesData.categoryDisplay.start, this.seriesData.seriesInfo.frequencyShort);
-      const end = this._helper.formatDate(this.seriesData.categoryDisplay.end, this.seriesData.seriesInfo.frequencyShort);
-      this.chart.setTitle({ text: '<b>' + this.seriesData.seriesInfo.displayName + '</b>' });
-      this.chart.setSubtitle({ text: 'Data Available From: ' + start + ' - ' + end, verticalAlign: 'middle', y: -20 });
-    }
-  }
-
-  checkPointCount(series0, point0, point1?, series1?) {
-    // Fiilter out null values
-    const pointCount = series0.points.filter(point => Number.isFinite(point.y));
-    // If only 1 non-null value exists, display data value as text
-    if (pointCount.length === 1) {
-      this.addSubtitle(point0, this.currentFreq, point1, series1);
-      series0.userOptions.states.hover.enabled = false;
-      series0.options.marker.states.hover.enabled = false;
-      point0.setState('');
-    }
-  }
-
-  addSubtitle(point0, freq, point1?, series1?) {
-    const decimals = this.seriesData.seriesInfo.decimals ? this.seriesData.seriesInfo.decimals : 1;
-    const unitsShort = this.seriesData.seriesInfo.unitsLabelShort;
-    const percent = this.seriesData.seriesInfo.percent;
-    const dateLabel = this.formatDateLabel(point0.x, freq);
-    let subtitleText = '';
-    subtitleText += Highcharts.numberFormat(point0.y, decimals, '.', ',') + '<br> (' + unitsShort + ') <br>';
-    subtitleText += series1 ?
-      this.formatTransformLabel(series1.name, percent) + '<br>' + Highcharts.numberFormat(point1.y, decimals, '.', ',') + '<br>' + dateLabel :
-      dateLabel;
-    this.chart.setSubtitle({
-      text: subtitleText,
-      verticalAlign: 'middle',
-      y: -20
-    });
-  }
-
-  formatTransformLabel(transformationName, percent) {
+  formatTransformLabel = (transformationName, percent) => {
     if (transformationName === 'c5ma') {
       return percent ? 'Annual Chg: ' : 'Annual % Chg: ';
     }
@@ -379,7 +358,7 @@ export class HighchartComponent implements OnInit, OnChanges {
     }
   }
 
-  formatDateLabel(date, freq) {
+  formatDateLabel = (date, freq) => {
     const year = Highcharts.dateFormat('%Y', date);
     if (freq === 'A') {
       return year;
@@ -403,17 +382,17 @@ export class HighchartComponent implements OnInit, OnChanges {
     }
   }
 
-  trimData(dataArray, start, end) {
-    const defaultRanges = this._helper.setDefaultCategoryRange(this.currentFreq, dataArray, this.defaultRange);
-    let startIndex = defaultRanges.start, endIndex = defaultRanges.end;
-    dataArray.forEach((item, index) => {
-      if (item[0] === start) {
-        startIndex = index;
-      }
-      if (item[0] === end) {
-        endIndex = index;
-      }
-    });
-    return dataArray.slice(startIndex, endIndex + 1);
+  noDataChart = (seriesData) => {
+    const title = seriesData.seriesInfo.displayName;
+    this.chartOptions.title = this.setChartTitle('<b>' + title + '</b><br>No Data Available');
+    this.chartOptions.exporting = { enabled: false };
+    this.chartOptions.legend = { enabled: false };
+    this.chartOptions.credits = { enabled: false };
+    this.chartOptions.yAxis = this.setYAxis();
+    this.chartOptions.xAxis = this.setXAxis();
+    this.chartOptions.series = [{
+      data: []
+    }];
+    this.chartOptions.lang = { noData: 'No Data Available' };
   }
 }
