@@ -4,7 +4,6 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { HttpModule } from '@angular/http';
-import { ChartModule } from 'angular2-highcharts';
 import { DataTableModule, SharedModule } from 'primeng/primeng';
 import { RecaptchaModule } from 'ng-recaptcha';
 import { AppComponent } from '../app.component';
@@ -28,6 +27,7 @@ import { AnalyzerHighstockComponent } from '../analyzer-highstock/analyzer-highs
 import { HelpComponent } from '../help/help.component';
 import { HelpDirective } from '../help.directive';
 import { ShareLinkComponent } from '../share-link/share-link.component';
+import { HighchartsChartModule } from 'highcharts-angular';
 
 @NgModule({
   declarations: [
@@ -56,7 +56,7 @@ import { ShareLinkComponent } from '../share-link/share-link.component';
   imports: [
     CommonModule,
     BrowserModule,
-    ChartModule,
+    HighchartsChartModule,
     FormsModule,
     ReactiveFormsModule,
     HttpModule,
